fix: ConnectWallet - tracked down bugs that caused thirdweb Tx to fail. - mayday situation
