+++ .gitignore
@@ -25,3 +25,4 @@ yarn-error.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts
+MINILEND_WHITEPAPER.txt
