+++ lib/thirdweb/client.ts
@@ -5,6 +5,7 @@ import { MINILEND_ADDRESS } from "../constants";
 
 const clientID = process.env.NEXT_PUBLIC_THIRDWEB_CLIENT_ID;
 
+
 if (!clientID) {
   throw new Error('Please add your clientID in environment variables');
 }
@@ -13,8 +14,8 @@ export const client = createThirdwebClient({
   clientId: clientID,
 });
 
-export const contract = getContract({ 
-  client,
-  chain: celo,
-  address: MINILEND_ADDRESS,
-});
+// export const contract = getContract({ 
+//   client,
+//   chain: celo,
+//   address: MINILEND_ADDRESS,
+// });
